Tavily Search
Tavily's Search is a search engine built specifically for AI agents (LLMs), delivering real-time, accurate, and factual results at speed.
In this step i have added the search engine agent file to the project. This file contains the search enginethat will be used to search for information on the web.